feat: 사진 직접 찍기, 갤러리에서 선택
